// src/api/ambulance-management/models/Dispatch.ts
export interface Dispatch {
    /**
     * 
     * @type {number}
     * @memberof Dispatch
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof Dispatch
     */
    incidentNumber: string;
    /**
     * 
     * @type {string}
     * @memberof Dispatch
     */
    callerName: string;
    /**
     * 
     * @type {string}
     * @memberof Dispatch
     */
    patientName: string;
    /**
     * 
     * @type {string}
     * @memberof Dispatch
     */
    streetAddress: string;
    /**
     * 
     * @type {string}
     * @memberof Dispatch
     */
    city: string;
    /**
     * 
     * @type {string}
     * @memberof Dispatch
     */
    dispatchReason: string;
    /**
     * 
     * @type {DispatchPriority}
     * @memberof Dispatch
     */
    priority: DispatchPriority;
    /**
     * 
     * @type {DispatchStatus}
     * @memberof Dispatch
     */
    status: DispatchStatus;
    /**
     * 
     * @type {string}
     * @memberof Dispatch
     */
    ambulanceCallSign: string;
    /**
     * 
     * @type {string}
     * @memberof Dispatch
     */
    destinationHospital: string;
    /**
     * 
     * @type {string}
     * @memberof Dispatch
     */
    dispatcherName: string;
    /**
     * 
     * @type {Date}
     * @memberof Dispatch
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof Dispatch
     */
    updatedAt: Date;
    /**
     * 
     * @type {string}
     * @memberof Dispatch
     */
    notes: string;
}