// src/api/ambulance-management/models/DispatchCreateRequest.ts
export interface DispatchCreateRequest {
    /**
     * 
     * @type {string}
     * @memberof DispatchCreateRequest
     */
    incidentNumber: string;
    /**
     * 
     * @type {string}
     * @memberof DispatchCreateRequest
     */
    callerName: string;
    /**
     * 
     * @type {string}
     * @memberof DispatchCreateRequest
     */
    patientName: string;
    /**
     * 
     * @type {string}
     * @memberof DispatchCreateRequest
     */
    streetAddress: string;
    /**
     * 
     * @type {string}
     * @memberof DispatchCreateRequest
     */
    city: string;
    /**
     * 
     * @type {string}
     * @memberof DispatchCreateRequest
     */
    dispatchReason: string;
    /**
     * 
     * @type {DispatchPriority}
     * @memberof DispatchCreateRequest
     */
    priority: DispatchPriority;
    /**
     * 
     * @type {DispatchStatus}
     * @memberof DispatchCreateRequest
     */
    status: DispatchStatus;
    /**
     * 
     * @type {string}
     * @memberof DispatchCreateRequest
     */
    ambulanceCallSign: string;
    /**
     * 
     * @type {string}
     * @memberof DispatchCreateRequest
     */
    destinationHospital: string;
    /**
     * 
     * @type {string}
     * @memberof DispatchCreateRequest
     */
    dispatcherName: string;
    /**
     * 
     * @type {Date}
     * @memberof DispatchCreateRequest
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof DispatchCreateRequest
     */
    notes?: string;
}