// src/api/ambulance-management/models/DispatchUpdateRequest.ts
export interface DispatchUpdateRequest {
    /**
     * 
     * @type {string}
     * @memberof DispatchUpdateRequest
     */
    incidentNumber: string;
    /**
     * 
     * @type {string}
     * @memberof DispatchUpdateRequest
     */
    callerName: string;
    /**
     * 
     * @type {string}
     * @memberof DispatchUpdateRequest
     */
    patientName: string;
    /**
     * 
     * @type {string}
     * @memberof DispatchUpdateRequest
     */
    streetAddress: string;
    /**
     * 
     * @type {string}
     * @memberof DispatchUpdateRequest
     */
    city: string;
    /**
     * 
     * @type {string}
     * @memberof DispatchUpdateRequest
     */
    dispatchReason: string;
    /**
     * 
     * @type {DispatchPriority}
     * @memberof DispatchUpdateRequest
     */
    priority: DispatchPriority;
    /**
     * 
     * @type {DispatchStatus}
     * @memberof DispatchUpdateRequest
     */
    status: DispatchStatus;
    /**
     * 
     * @type {string}
     * @memberof DispatchUpdateRequest
     */
    ambulanceCallSign: string;
    /**
     * 
     * @type {string}
     * @memberof DispatchUpdateRequest
     */
    destinationHospital: string;
    /**
     * 
     * @type {string}
     * @memberof DispatchUpdateRequest
     */
    dispatcherName: string;
    /**
     * 
     * @type {Date}
     * @memberof DispatchUpdateRequest
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof DispatchUpdateRequest
     */
    notes?: string;
}